Added lesson10 and test
